refactor(cli): split version functions into those that display and those that do not
