+++ packages/aws-cdk/lib/cli/cli.ts
@@ -12,7 +12,6 @@ import { prettyPrintError } from './pretty-print-error';
 import { GLOBAL_PLUGIN_HOST } from './singleton-plugin-host';
 import type { Command } from './user-configuration';
 import { Configuration } from './user-configuration';
-import * as version from './version';
 import { asIoHelper } from '../../lib/api-private';
 import type { IReadLock } from '../api';
 import { ToolkitInfo, Notices } from '../api';
@@ -30,6 +29,8 @@ import { getMigrateScanType } from '../commands/migrate';
 import { execProgram, CloudExecutable } from '../cxapp';
 import type { StackSelector, Synthesizer } from '../cxapp';
 import { ProxyAgentProvider } from './proxy-agent';
+import { isDeveloperBuildVersion, version, versionNumber } from './version';
+import { displayVersionMessage } from './display-version';
 
 if (!process.stdout.isTTY) {
   // Disable chalk color highlighting
@@ -78,7 +79,7 @@ export async function exec(args: string[], synthesizer?: Synthesizer): Promise<n
     await ioHost.defaults.debug(`Error while checking for platform warnings: ${e}`);
   }
 
-  await ioHost.defaults.debug('CDK Toolkit CLI version:', version.displayVersion());
+  await ioHost.defaults.debug('CDK Toolkit CLI version:', version());
   await ioHost.defaults.debug('Command line arguments:', argv);
 
   const configuration = await Configuration.fromArgsAndFiles(ioHelper,
@@ -103,7 +104,7 @@ export async function exec(args: string[], synthesizer?: Synthesizer): Promise<n
     context: configuration.context,
     output: configuration.settings.get(['outdir']),
     httpOptions: { agent: proxyAgent },
-    cliVersion: version.versionNumber(),
+    cliVersion: versionNumber(),
   });
   const refreshNotices = (async () => {
     // the cdk notices command has it's own refresh
@@ -164,7 +165,7 @@ export async function exec(args: string[], synthesizer?: Synthesizer): Promise<n
     await outDirLock?.release();
 
     // Do PSAs here
-    await version.displayVersionMessage(ioHelper);
+    await displayVersionMessage(ioHelper);
 
     await refreshNotices;
     if (cmd === 'notices') {
@@ -499,7 +500,7 @@ export async function exec(args: string[], synthesizer?: Synthesizer): Promise<n
         });
       case 'version':
         ioHost.currentAction = 'version';
-        return ioHost.defaults.result(version.displayVersion());
+        return ioHost.defaults.result(version());
 
       default:
         throw new ToolkitError('Unknown command: ' + command);
@@ -637,7 +638,7 @@ export function cli(args: string[] = process.argv.slice(2)) {
     .catch((err) => {
       // Log the stack trace if we're on a developer workstation. Otherwise this will be into a minified
       // file and the printed code line and stack trace are huge and useless.
-      prettyPrintError(err, version.isDeveloperBuild());
+      prettyPrintError(err, isDeveloperBuildVersion());
       process.exitCode = 1;
     });
 }
