+++ packages/aws-cdk/lib/cli/util/yargs-helpers.ts
@@ -31,7 +31,7 @@ export function yargsNegativeAlias<T extends { [x in S | L]: boolean | undefined
  * @returns the current version of the CLI
  */
 export function cliVersion(): string {
-  return version.displayVersion();
+  return version.version();
 }
 
 /**
