+++ packages/aws-cdk/lib/commands/context.ts
@@ -5,7 +5,7 @@ import type { Context } from '../api/context';
 import type { IoHelper } from '../api-private';
 import { renderTable } from '../cli/tables';
 import { PROJECT_CONFIG, PROJECT_CONTEXT, USER_DEFAULTS } from '../cli/user-configuration';
-import * as version from '../cli/version';
+import { displayVersionMessage } from '../cli/display-version';
 
 /**
  * Options for the context command
@@ -71,7 +71,7 @@ export async function contextHandler(options: ContextOptions): Promise<number> {
       await listContext(ioHelper, options.context);
     }
   }
-  await version.displayVersionMessage(ioHelper);
+  await displayVersionMessage(ioHelper);
 
   return 0;
 }
