+++ packages/aws-cdk/test/cli/version.test.ts
@@ -6,8 +6,9 @@ import * as sinon from 'sinon';
 import { setTimeout as _setTimeout } from 'timers';
 import { promisify } from 'util';
 import * as npm from '../../lib/cli/util/npm';
-import { displayVersionMessage, getVersionMessages, isDeveloperBuild, VersionCheckTTL } from '../../lib/cli/version';
+import { isDeveloperBuildVersion } from '../../lib/cli/version';
 import { TestIoHost } from '../_helpers/io-host';
+import { displayVersionMessage, getVersionMessages, VersionCheckTTL } from '../../lib/cli/display-version';
 
 jest.setTimeout(10_000);
 
@@ -150,7 +151,7 @@ test('isDeveloperBuild call does not throw an error', () => {
   // that the value is `true` when running tests, because I won't want to make too
   // many assumptions for no good reason.
 
-  isDeveloperBuild();
+  isDeveloperBuildVersion();
 
   // THEN: should not explode
 });
